[FEAT] 채널

+++ ChattingServer/src/main/java/com/teambind/auth/dto/Channel.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public record Channel(
 		ChannelId channelId,
 		String title,
-		int headCount
+		Long headCount
 ) {
 	@Override
 	public boolean equals(Object o) {
